Convert to alpine image
Removed unneeded files from fork. This is built purely for my purposes, running on a AMD64 server as an alpine image. Arm builds are not needed for my purposes. Refactored entrypoint script to search for correct files and replace if not found - original script wasn't working correctly for some reason.